chore: moved URL creation logic to UrlConfig
